requirements.txt
anthropic>=0.39.0
elevenlabs>=1.0.0
trafilatura>=1.12.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
lxml_html_clean>=0.4.0

pydub>=0.25.1
google-cloud-texttospeech>=2.17.0

# Fase 2: Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
python-multipart>=0.0.9
